Index VotingLists instead of Votes records
